+++ dev/src/clasew/sandbox.clj
@@ -8,3 +8,4 @@
 
 ;; Demonstrate record coercion
 (def p pprint)
+